Tried to resolve ontologies in directories.
Some BioPortal ontologies come as ZIP files with multiple OWL files. To
find the main file, it was tried to find the single file that is not
imported by any other file. Unfortunately, this approach fails with ICPS
which has multiple files that are not referenced any other file. Thus,
the approach will be removed in future commits.