Resolves #2, #4.
Also updated to the OWL API 5.1.0 which uses new parsers. That possibly
also solves the UMLS issues.
Very importantly, added an AutoIRI Mapper for ontologies stored in
directories in order to actually use the delivered files and not ignore
them.
At the moment, there are only two ontologies that actually have multiple
files, ICPS and OCRE. For those, their main file is identified by
comparison to the download file name which works in both cases. If, in
the future, new ontologies with multiple files are added, the
OntologyLoader would have to add adequate strategies.